utils: move uuid serde tests into utils dir
This fixes dependencies in which serde is needed towards the root of the
dependency tree, and utils much deeper. since the test was for uuid
which lives in utils, importing created a circulr dependency.
